Fix src/server/redis_connection.cc

+++ src/server/redis_connection.cc
@@ -196,7 +196,7 @@ void Connection::UnSubscribeChannel(const std::string &channel) {
 
 void Connection::UnSubscribeAll(const unsubscribe_callback &reply) {
   if (subscribe_channels_.empty()) {
-    if (reply != nullptr) reply("", subcribe_patterns_.size());
+    if (reply != nullptr) reply("", static_cast<int>(subcribe_patterns_.size()));
     return;
   }
   int removed = 0;
@@ -233,7 +233,7 @@ void Connection::PUnSubscribeChannel(const std::string &pattern) {
 
 void Connection::PUnSubscribeAll(const unsubscribe_callback &reply) {
   if (subcribe_patterns_.empty()) {
-    if (reply != nullptr) reply("", subscribe_channels_.size());
+    if (reply != nullptr) reply("", static_cast<int>(subscribe_channels_.size()));
     return;
   }
